requirements.txt
pandas
openpyxl
gdown
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
google-cloud-storage
ipykernel
google-cloud-vision
SQLAlchemy
scikit-learn